feat: Add offline token creation test (#26771)
This commit adds an offline token creation test. We did not have this as
part of the 3.4 test in the interest of time. This adds automated
testing to make sure the flow we expect happens which includes:

- Ability to create admin tokens via the CLI
- Token creation on startup
- Database creation on start

With this we can be more confident that the feature works as expected
and that other changes that might affect it have regression coverage.