feat: Add experimental multi file plugin support
This commit does a few things:

1. Adds the ability to have plugins with multiple files
2. Adds the ability for updating plugins and configuration files
3. Adds the ability to query for information about plugins with both
   APIs and System Tables
4. Allows plugins to work with multiple files making code much more
   modular
5. Allows adding non python files to a plugin for use in ones code. f.ex
   you could have a JSON configuration file included with the plugin that
   you can then update later, have the plugin automatically reload, and
   be all set with your changes.

While this feature is in a fairly usable state it still needs some
polishing for the rough edges, including things like being able to
upload a plugin to the server as long as one has admin access. In order
to use this feature you do so like any other plugin and drop the files
in the plugin folder that you point influxdb3 at and set your trigger for
it via the CLI.

The difference here is that you use the new --path argument instead of
--plugin-file. For a multifile plugin place a directory of files inside
the plugin dir with a __main__.py file and use the directory name with
the --path argument when creating a trigger to enable it. The server will
take care of the rest and your code should work as you'd expect it too.