fix(tokio): futures::executor::block_on is not compatible with tokio (#26803)
* fix(tokio): futures::executor::block_on is not compatible with tokio

futures::executor::block_on takes over the thread that calls it with a
while loop expecting another thread or runtime to drive the contained
future to completion. This doesn't work with tokio's runtime thread
expectations where it expects to swap tasks off threads as it likes.
While it doesn't always cause an issue with tokio, the futures block_on
can stall the tokio runtime.

The method tokio::task::block_in_place runs the contained closure in a
blocking manner without blocking the runtime. Within it then, you can
use Handle::current().block_on to enter an async context to execute
futures correctly on tokio's runtime.

These tokio versions are not compatible with the "current_thread"
runtime option of tokio.

This commit also includes the uri path when an authentical request to
the server's http api fails. This is useful for understand the target of
the unauthorized request.

* fix: remove current thread tokio runtime type

It is not compatible with our execution model.

Co-authored-by: Phil Bracikowski <13472206+philjb@users.noreply.github.com>