fix(reauthing object store): respect expiry in the file (#1342)
This adjusts the background reloading task for object store credentials.
It would ignore the expiry in the file if the credentials weren't
changing but this might mean that a check at the expiry time is missed.

This pr also makes sure the temporary runtime is shutdown when building
with the no_license feature enabled, which is the standard build config
for our hosting partner.

* fixes #1341