chore: cherry-pick recent enterprise bug fixes (#26813)
* fix: reject lp lines containing duplicate fields

The main problem this PR resolves is that ingesters currently accept line
protocol containing duplicates of the same field. This results in the WAL
containing N times the number of values for the duplicated fields where N is
the number of duplicates.

This causes the server to crash after the WAL files are persisted to object
store during the gen1 file snapshotting process. Since the corrupted WAL is
persisted to object store, the server will again crash next time it attempts to
start up during WAL replay & snapshotting.

The fix here is to introduce a `HashSet<ColumnId>` that we use when qualifying
each line in a batch that lets us detect a duplicate field. If we detect a
duplicate field, we return a `WriteLineError` indicating the problematic field.
This solution slightly increases the memory usage in the `WriteValidator`, but
we make a best effort to avoid this by re-using the same allocation for all
lines in a batch. The upper bound on the extra memory allocated for this on a
per-batch basis is 2**16*16 = 1MB, but should typically be much smaller since
it scales with the number of fileds on a given line.

* fix(reauthing object store): respect expiry in the file

This adjusts the background reloading task for object store credentials.
It would ignore the expiry in the file if the credentials weren't
changing but this might mean that a check at the expiry time is missed.

This pr also makes sure the temporary runtime is shutdown when building
with the no_license feature enabled, which is the standard build config
for our hosting partner.

* fixes #1341

---------

Co-authored-by: Phil Bracikowski <13472206+philjb@users.noreply.github.com>