refactor: finish table index updates before attempting to create new indices (#26838)
This PR attempts to address https://github.com/influxdata/influxdb/issues/26810

The working theory is that there is a potential race condition between table
index updates and new table index creation where it's possible that table index
snapshots are deleted during the background update tasks while the spawning
context continues to attempt to list all table index snapshots (done for the
sake of detecting new tables that don't yet exist in the cache).

This PR adjusts the sequencing of these such that they are non-overlapping --
we now wait for all index updates to finish before attempting to list snapshots
and create any missing indices.

Unfortunately we haven't yet been able to reproduce the server freeze reported
in #26810, so we don't know for sure that this fixes the issue. But it at least
cleans the code up a little.