chore: backport stop node only if node exists fix
This backports changes from https://github.com/influxdata/influxdb_pro/pull/1394
that prevents errors when trying to stop a node that doesn't exist in the catalog.

In local testing when the backing storage gets wiped out before node gracefully
shuts down then it leaves the `StopNodeLog` written to the store. This means in
subsequent start, it runs into error trying to stop a node that's not been
registered. This commit stops node only if it already exists.