feat: Add plugin file system table (#26875)
This commit adds a system table that can be queried in order to show all
of the files for a plugin via a system table under _internal. We also
add some tests to make sure the tables contain what we would expect if
we delete plugins, add plugins, or have no plugins set.

This is part of the multi file plugin updates and lays the foundation
for being able to test that those changes are working.