feat: Add show plugins command (#26878)
This command exposes the ability to query the system table that was
added in f7021ff6 via the CLI making it so that operators can easily
see what files exist for plugins loaded into influxdb3. In the future
this will list multiple files for plugins that are more than a singular
file in size.

For now we expose the table in much the same way as we have done before.
Since the table is an internal table this does require an admin token to
view the table.

Tests were also included so as to avoid regressions with this command.