feat: Allow updating plugin files for a trigger (#26892)
This commit adds the ability to update the files for a given trigger at
runtime by uploading new code to the server. This is done via an admin
only command so that not just anyone can upload code to the server.
We look for the file associated with the plugin and update it on the
server side with the content sent to it via an HTTP request.