feat: Add multifile plugin support to InfluxDB 3
This commit adds multifile plugin support for InfluxDB 3 and allows
users to write more complex plugins that have more than 1 file. The
following commit adds tests and the following abilities:

1. Multifile plugins can be created just like normal single file plugins
2. They can be uploaded as well just like single file plugins
3. The new path argument handles dirs and single files for update and
   creation of plugins

With this we expect users to be able to create more complex plugins and
to be able to update them easily!