feat: Allow uploading files on trigger creation (#26895)
This allows users to create triggers while uploading their code via
a new --upload flag. This is an admin only feature as uploading
arbitrary code is something that only the person in control of the
server should be able to do. Much like before you create the trigger
the same way except now you add the --upload flag. This then uses the
local path on your computer rather than the one in the plugin directory
and creates the file in the plugin directory on the server before
creating the trigger.