chore(log_filter; help): reduce query logging; update help text
Several crates log an incoming query at the info! level. This was
producing many lines of logs (5+) for a single query. It is quite
verbose especially since each was also including the query string
itself.

This changeset updates the default log filter to remove some of the
excess query logging. The help text has been updated to match and
formatting applied.

* this is a backport of influxdata/influxdb_pro#1514