feat(http,write): support v1 db/rp style names in all write apis
This changes the validation of database names to support v1 db/rp style
names in all apis. Only one forward slash is allowed. Tests, errors, and
logging are updated. Validation of db name length is pulled into
monolith code so we can return a 400 instead of a 500 which the iox
NamespaceName validation was causing.

* this is a port of influxdata/influxdb_pro#1516