fix(write precision): support short forms and truncate server time in write_lp (#26926)
* fix(write precision): support short forms in v3 write_lp endpoint

The v3 write_lp endpoint that core and enterprise exposes should support
the short forms of the precision arguments: s, ms, us, u, ns, and n. These
forms are supported by v3 cloud and short forms by v2 and v1. Tests are
added to ensure from_str/parse match the deserialization methods.

* this is a backport of influxdata/influxdb_pro#1527

* feat(write): truncate server-provided timestamps to precision

v1 and v2 truncate the ingest/server provided timestamps according to
the precision. v3 needs to match v1/v2 behavior and it is a
performance opportunity for those who want the db server to determine
the timestamp of incoming data with a precision less granular than
nanoseconds.

This pr updates the handling of server timestamps for the /write,
/api/v2/write, and /api/v3/write_lp endpoints to truncate. It
additionally checks for overflow now when an explicit timestamp is
provided with a precision. Parquet and pacha each have a write path that
duplicated some of the precision handling code; that has been
consolidated into methods on Precision. Tests have been added for the
original and new functionality. The overflow creates a new possible
error that was silent before.

* this is a backport of influxdata/influxdb_pro#1533