feat(clap): return an error if enterprise arg cluster-id is used
It has been a common mistake to run core's binary thinking it is
enterprise 3. This seems to error first on the --cluster-id arg which
is present in enterprise 3 but not core. Clap gives an "unexpected
argument" error which is confusing as the user thought they were running
enterprise.

This commit updates Core to look for this argument and always return
an error explaing that this is core and an enterprise argument was used.
This gives clear direction for the problem and how to fix it.

* closes #26967