feat: add TLS certificate reloading on SIGHUP
Add TLS certificate reloading on SIGHUP. For httpd service certificates,
the configuration is reloaded and certificate and key file locations are
updated accordingly. For opentsdb service certificates, certificates and
keys at the existing locations are reloaded. If reloading a certificate
fails, the currently loaded certificate continues to be used.

Also adds file permission checking for TLS certificates and private
keys.