feat(port): port from ent to core (#26996)
* feat(port): port from ent to core

This ports a number of changes from enterprise to core. Primarily this
adds a new telemetry field for how influxdb3 has been started (aka
invocation method). Most of the changes are to sync core code with the
shape of ent code to make future ports easier. There are some error
reporting improvements and logging improvements mixed in there too.

* follows #26991

* refactor: copilot pr feedback

---------

Co-authored-by: Trevor Hilton <thilton@influxdata.com>