feat: Bubble error up to writer if fields are dropped  (#26998)
Modify `WritePoints` to return an error if a user defined `time` field is inserted. Previously we would silently prune it, this PR will bubble an error up to the writer to inform that a field was dropped but not the entire point.

(cherry picked from commit 362217b3c142444ceaf115fbd7fec66215adbc0f)