feat: add TLS certificate reloading on SIGHUP (#26994)
* feat: add TLS certificate reloading on SIGHUP

Add TLS certificate reloading on SIGHUP. For httpd service certificates,
the configuration is reloaded and certificate and key file locations are
updated accordingly. For opentsdb service certificates, certificates and
keys at the existing locations are reloaded. If reloading a certificate
fails, the currently loaded certificate continues to be used.

Also adds file permission checking for TLS certificates and private
keys.

* chore: rewrite some tests to get certificates without tricks

Rewrite some tests so that tricks are not used to get a hold of the TLS
certificate in-use. Instead, an HTTP client is used to get the
certificate presented by the http service.

* fix: correct race conditions in httpd.Service

- Correct race conditions present in httpd.Service.
- Improve error messages.
- Rename some methods for clarity.
- Fix issues in tests.
- Fix comments.